+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Reorganize Hutch config - instead of setting config values when connecting to Hutch when using publisher, do it right after the initialization.
 - Make params filter customizable for the logger
+- Remove sensitive info when creating distributed trace records.
 
 ## 0.2.0
 - Implement "Safe Producer" extension allowing to retry delivery later in an automated way when publishing of the message fails.